Don't replace '_' with '-' in framework names
Earlier the Kotlin/Native compiler used the output framework name as
a prefix for symbols of this framework. Such an approach caused
compilation error when this name contains the '-' symbol. To
workaround it on the plugin side we replaced dashes with underscores.
Currently the compiler correctly handles dashes in framework names
and there is no need to mangle them in the plugin.

GitHub issue: https://github.com/JetBrains/kotlin-native/issues/2588